Move firebase modules to folder
